Refactor: Modularize MCP server logic into separate files
